+++ src/pages/users/signup.tsx
@@ -4,7 +4,7 @@ import { CircularProgress } from "@material-ui/core";
 
 import UserForm from "../../components/Users/Form";
 import Messages from "../../utils/messages";
-import { useCurrentUser } from "../../hooks";
+import { useAllUsers, useCurrentUser } from "../../hooks";
 import { SERVER_INVITE_BY_TOKEN } from "../../apollo/client/queries";
 import { noCache } from "../../utils/apollo";
 import { useLazyQuery } from "@apollo/client";
@@ -13,6 +13,7 @@ import { redeemedInviteToken } from "../../utils/invite";
 const SignUp = () => {
   const token = redeemedInviteToken();
   const currentUser = useCurrentUser();
+  const [users, _setUsers, usersLoading] = useAllUsers();
   const [invite, setInvite] = useState<ServerInvite>();
   const [getInviteRes, inviteRes] = useLazyQuery(
     SERVER_INVITE_BY_TOKEN,
@@ -37,9 +38,9 @@ const SignUp = () => {
     if (currentUser?.isAuthenticated) Router.push("/");
   }, [currentUser]);
 
-  if (inviteRes.loading) return <CircularProgress />;
+  if (inviteRes.loading || usersLoading) return <CircularProgress />;
   if (currentUser) return <>{Messages.users.alreadyRegistered()}</>;
-  if (invite) return <UserForm />;
+  if (invite || users.length === 0) return <UserForm />;
   return <>{Messages.invites.redeem.inviteRequired()}</>;
 };
